Fixed the x86_64 setjmp implementation.
* typedef for jmp_buf was using int where it should be long.
* setjmp was clearing the buffer pointer rather than the signal mask before
  calling sigsetjmp.
* KDL now works without crashing on x86_64.

+++ headers/posix/arch/x86_64/arch_setjmp.h
@@ -6,6 +6,6 @@
 #define _ARCH_SETJMP_H_
 
 
-typedef int __jmp_buf[8];
+typedef unsigned long __jmp_buf[8];
 
 #endif	/* _ARCH_SETJMP_H_ */
